v1.1.10; Issue #9
First attempt at solving issue #9 (Make it possible to search "all" Repository search topics). When "all" is selected, the code searches for a match in any of the columns that are shown on the Shiny App (note: the code does not search for ALL columns in the df). Also corrected a small typo in the disclaimer text.