BK: Uploading preprints scraping script
The current version of the script scrapes preprints from OSF. It searches for ESM, EMA, and daily diary mentions (see keywords on line 67) in titles, abstracts, and tags (see selection on lines 43-47). Then, it downloads the found preprints, for which it requires having a folder "downloaded preprints" in the user's working directory.
Failing to fetch preprint details (error message on lines 104-106) does not seem to cause any issues.